FIX: OBOE
The consequences of this off-by-one error were especially hard to
catch as a NaN in post_trial_gap lead to -1 in both timing_stim and
timing_response leading to unexpected behaviour (experiment stopped)
in jspsych-poldrack-single-stim.js.

Not the best fix but it'll have to stay as tech debt for now.
with '#' will be ignored, and an empty message aborts the commit.  #
On branch master # Your branch is up-to-date with 'origin/master'.  #